Revert "Revert "Popravljeno racunanje PDV kod razlicitih pdv-a artikla, popravljen error prilikom otvaranja racuna u novoj godini""
This reverts commit a32ea0036fee15b70ef4975a0253e59e678566fa.

+++ Poslovni/Klase/Artikl.cs
@@ -11,7 +11,7 @@ namespace Poslovni
 {
     public struct Artikl
     {
-        public int sifra;
+        public long sifra;
         public string naziv; 
         public int kolicina;
         public int poreza_grupa;
